refactor(memory): split memory system ownership
